pyproject.toml
[tool.poetry]
name = "my-python-app"
version = "0.1.0"
description = "A template for Python applications using Poetry"
authors = ["Ofelia Webb <ofelia.b.webb@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13"
transformers = "^4.38.2"
pdfplumber = "^0.10.3"
python-docx = "^1.1.0"
torch = "^2.2.1"
scikit-learn = "^1.4.1.post1"
joblib = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"

[tool.poetry.scripts]
my-app = "app.main:main"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"